{{cookiecutter.project_name}}/requirements.txt
ipython
ipdb
flake8
Django
psycopg2
https://github.com/jezdez/django-configurations/archive/master.tar.gz
raven
WebTest
django-webtest
model-mommy
django-debug-toolbar
django-model-utils
django-redis-cache
hiredis
django-bootstrap3
django-vanilla-views
django-extensions
django-clear-cache
pyreact
django-pipeline
djangorestframework
django-filter
